Move e2e tests out of stupid dirs
