+++ cypress/e2e/export-room-members.spec.ts
@@ -1,7 +1,7 @@
 /// <reference types="cypress" />
 
-import RoomUtils from "../../utils/room-utils";
-import RandomUtils from "../../utils/random-utils";
+import RoomUtils from "../utils/room-utils";
+import RandomUtils from "../utils/random-utils";
 import { normalize } from "matrix-js-sdk/src/utils";
 
 describe("Export room members feature", () => {
